av1/requirements.txt
pymongo
transformers
torch
google-generativeai